movies_admin/pyproject.toml
[tool.poetry]
name = "movies-admin"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.11"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
django-split-settings = "^1.3.2"
psycopg = {extras = ["c"], version = "^3.2.3"}
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
